+++ Runtime/ConnectionStatus.cs
@@ -1,0 +1,8 @@
+public enum ConnectionStatus
+{
+    ConnectionPending,
+    Connected,
+    DisconnectionPending,
+    Disconnected,
+    Error
+}