+++ Runtime/TwitchIRC.cs
@@ -1,19 +1,14 @@
 ﻿using System.Collections;
-using System.Collections.Concurrent;
-using System.Net.Sockets;
-
-using System.Threading;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace Incredulous.Twitch
 {
     /// <summary>
-    /// A class which represents a connection to the Twitch IRC API.
+    /// A class which manages connections to the Twitch IRC API.
     /// </summary>
     public partial class TwitchIRC : MonoBehaviour
     {
-        #region Inspector Values
-
         [Header("Server Information")]
 
         /// <summary>
@@ -29,7 +24,7 @@ namespace Incredulous.Twitch
         /// <summary>
         /// Twitch login details.
         /// </summary>
-        public TwitchDetails twitchDetails;
+        public TwitchCredentials twitchCredentials;
 
 
         [Header("Read Information")]
@@ -56,17 +51,81 @@ namespace Incredulous.Twitch
         [Header("Other Settings")]
 
         /// <summary>
-        /// General settings for the Twitch client.
+        /// Whether a connection to Twitch should be established on Start.
         /// </summary>
-        public Settings settings;
-
-
-        [Header("Client Chatter Object")]
+        public bool connectOnStart = true;
 
         /// <summary>
-        /// Contains some information about the client user (OAuth).
+        /// Whether all IRC messages should be logged to the debug console.
         /// </summary>
-        public Chatter clientChatter;
+        public bool debugIRC = true;
+
+
+        /// <summary>
+        /// The client user's Twitch tags.
+        /// </summary>
+        public IRCTags clientUserTags { get; private set; }
+
+        /// <summary>
+        /// The current connection status.
+        /// </summary>
+        public ConnectionStatus status => connection?.status ?? ConnectionStatus.Disconnected;
+
+
+        /// <summary>
+        /// The current Twitch connection.
+        /// </summary>
+        private TwitchConnection connection;
+
+        /// <summary>
+        /// A queue for connection alerts.
+        /// </summary>
+        private Queue<ConnectionAlert> alertQueue = new Queue<ConnectionAlert>();
+
+        #region Unity MonoBehaviour Messages
+
+        private void Start()
+        {
+            if (connectOnStart)
+                StartCoroutine(ConnectCoroutine());
+        }
+
+        private void Update()
+        {
+            if (connection == null)
+                return;
+
+            while (!connection.chatterQueue.IsEmpty)
+            {
+                if (connection.chatterQueue.TryDequeue(out var chatter))
+                    ChatMessageEvent?.Invoke(chatter);
+            }
+
+            while (!connection.connectionAlertQueue.IsEmpty)
+            {
+                if (connection.connectionAlertQueue.TryDequeue(out var alert))
+                    HandleConnectionAlert(alert);
+            }
+
+            while (alertQueue.Count > 0)
+            {
+                var alert = alertQueue.Dequeue();
+                HandleConnectionAlert(alert);
+            }
+
+            if (clientUserTags != connection.clientUserTags)
+                clientUserTags = connection.clientUserTags;
+        }
+
+        private void OnDestroy()
+        {
+            BlockingDisconnect(connection);
+        }
+
+        private void OnDisable()
+        {
+            BlockingDisconnect(connection);
+        }
 
         #endregion
 
@@ -80,7 +139,7 @@ namespace Incredulous.Twitch
         /// <summary>
         /// A delegate which handles a status update from the Twitch IRC client.
         /// </summary>
-        public delegate void StatusUpdateEventHandler(StatusType statusType, string message, int percent);
+        public delegate void ConnectionAlertEventHandler(ConnectionAlert connectionAlert);
 
         /// <summary>
         /// An event which is triggered when a new chat message is received.
@@ -90,71 +149,7 @@ namespace Incredulous.Twitch
         /// <summary>
         /// An event which is triggered when the connection status changes.
         /// </summary>
-        public event StatusUpdateEventHandler StatusUpdateEvent;
-
-        #endregion
-
-        /// <summary>
-        /// The number of milliseconds Twitch requires between IRC writes.
-        /// </summary>
-        private const int twitchRateLimitSleepTime = 1750;
-
-        /// <summary>
-        /// The TCP client.
-        /// </summary>
-        private TcpClient client;
-
-        /// <summary>
-        /// The thread which handles sending to the IRC server.
-        /// </summary>
-        private Thread sendThread = null;
-
-        /// <summary>
-        /// The thread which handles receiving from the IRC server.
-        /// </summary>
-        private Thread receiveThread = null;
-
-        /// <summary>
-        /// A thread safe queue which performs actions on the main thread.
-        /// </summary>
-        private ConcurrentQueue<System.Action> taskQueue = new ConcurrentQueue<System.Action>();
-
-        /// <summary>
-        /// Whether the client should currently be connected to Twitch.
-        /// </summary>
-        private bool connected
-        {
-            get => _connected == 1;
-            set => Interlocked.Exchange(ref _connected, value ? 1 : 0);
-        }
-        private int _connected = 0;
-
-        #region Unity MonoBehaviour functions
-
-        private void Start()
-        {
-            if (settings.connectOnStart)
-                StartCoroutine(ConnectCoroutine());
-        }
-
-        private void Update()
-        {
-            while (taskQueue.Count > 0)
-            {
-                if (taskQueue.TryDequeue(out var task))
-                    task.Invoke();
-            }
-        }
-
-        private void OnDestroy()
-        {
-            BlockingDisconnect();
-        }
-
-        private void OnDisable()
-        {
-            BlockingDisconnect();
-        }
+        public event ConnectionAlertEventHandler ConnectionAlertEvent;
 
         #endregion
 
@@ -173,16 +168,7 @@ namespace Incredulous.Twitch
         [ContextMenu("Disconnect IRC")]
         public void Disconnect()
         {
-            StartCoroutine(DisconnectCoroutine());
-        }
-
-        /// <summary>
-        /// Ping the Twitch server.
-        /// </summary>
-        [ContextMenu("Send PING")]
-        private void SendPing()
-        {
-            SendCommand("PING :tmi.twitch.tv", true);
+            StartCoroutine(DisconnectCoroutine(connection));
         }
 
         /// <summary>
@@ -191,81 +177,57 @@ namespace Incredulous.Twitch
         private IEnumerator ConnectCoroutine()
         {
             // End any current connection
-            if (connected)
-                yield return StartCoroutine(DisconnectCoroutine());
-
-            // Wait for previous threads to close (if there are any)
-            while (receiveThread != null && receiveThread.IsAlive)
-                yield return null;
-            while (sendThread != null && sendThread.IsAlive)
-                yield return null;
+            StartCoroutine(DisconnectCoroutine(connection));
 
             // Verify that login information has been provided
-            if (twitchDetails.oauth.Length <= 0 || twitchDetails.nick.Length <= 0 || twitchDetails.channel.Length <= 0)
+            if (twitchCredentials.oauth.Length <= 0 || twitchCredentials.username.Length <= 0 || twitchCredentials.channel.Length <= 0)
             {
-                ConnectionStateAlert(StatusType.Error, "Missing required details! Check your Twitch details.");
+                ConnectionAlertEvent?.Invoke(ConnectionAlert.MissingLogin);
                 yield break;
             }
 
             // Fix formatting (twitchapps.com)
-            if (twitchDetails.oauth.StartsWith("oauth:"))
-                twitchDetails.oauth = twitchDetails.oauth.Substring(6);
+            if (twitchCredentials.oauth.StartsWith("oauth:"))
+                twitchCredentials.oauth = twitchCredentials.oauth.Substring(6);
 
-            // Connect to Twitch IRC
-            client = new TcpClient(ircAddress, port);
+            // Create a new connection to Twitch IRC
+            connection = new TwitchConnection(this);
 
-            if (!client.Connected)
+            // Check the connection
+            if (!connection.tcpClient.Connected)
             {
-                ConnectionStateAlert(StatusType.Error, "Failed connecting to Twitch IRC.");
+                ConnectionAlertEvent?.Invoke(ConnectionAlert.NoConnection);
                 yield break;
             }
-            connected = true;
 
-            // Clear the task queue (NOTE: ConcurrentQueue does not contain a Clear method in .NET Standard 2.0)
-            while (!taskQueue.IsEmpty)
-                taskQueue.TryDequeue(out _);
-
-            // Clear the output queue
-            while (!outputQueue.IsEmpty)
-                outputQueue.TryDequeue(out _);
-
-            // Initialize threads
-            receiveThread = new Thread(() => IRCInputProc());
-            sendThread = new Thread(() => IRCOutputProc());
-            connectionThread = new Thread(() => CheckSocketProc());
-
-            // Start threads
-            receiveThread.Start();
-            sendThread.Start();
-            connectionThread.Start();
-
-            // Queue login commands
-            SendCommand("PASS oauth:" + twitchDetails.oauth.ToLower(), true);
-            SendCommand("NICK " + twitchDetails.nick.ToLower(), true);
-            SendCommand("CAP REQ :twitch.tv/tags twitch.tv/commands", true);
+            // Begin the threads and attempt to authenticate
+            connection.Begin();
         }
 
         /// <summary>
         /// A coroutine which instructs the send/receive threads to terminate, waits for them to end, then closes the TCP connection.
         /// </summary>
-        private IEnumerator DisconnectCoroutine()
+        private IEnumerator DisconnectCoroutine(TwitchConnection connection)
         {
-            if (!connected)
+            if (connection == null)
                 yield break;
 
-            // Instruct the threads to stop running
-            connected = false;
+            if (connection.status == ConnectionStatus.Disconnected || connection.status == ConnectionStatus.DisconnectionPending)
+                yield break;
+
+            // Stop the connection
+            connection.End();
 
             // Wait for threads to close
-            while (receiveThread.IsAlive)
-                yield return null;
-            while (sendThread.IsAlive)
-                yield return null;
-            while (connectionThread.IsAlive)
+            while (connection.threadsActive)
                 yield return null;
 
             // Close the TcpClient
-            client.Close();
+            connection.Close();
+
+            // Reset connection
+            if (this.connection == connection)
+                this.connection = null;
 
             Debug.LogWarning("Disconnected from Twitch IRC");
         }
@@ -273,142 +235,42 @@ namespace Incredulous.Twitch
         /// <summary>
         /// Disconnect from Twitch IRC. <b>Blocks the main thread. Use carefully.</b>
         /// </summary>
-        private void BlockingDisconnect()
+        private void BlockingDisconnect(TwitchConnection connection)
         {
-            if (!connected)
+            if (connection == null)
                 return;
 
-            // Instruct the threads to stop running
-            connected = false;
+            if (connection.status == ConnectionStatus.Disconnected || connection.status == ConnectionStatus.DisconnectionPending)
+                return;
 
-            // Wait for threads to close
-            receiveThread.Join();
-            sendThread.Join();
-            connectionThread.Join();
+            // End the connection
+            connection.BlockingEndAndClose();
 
-            // Close the TcpClient
-            client.Close();
+            // Reset connection
+            if (connection == this.connection)
+                this.connection = null;
 
             Debug.LogWarning("Disconnected from Twitch IRC");
         }
 
         /// <summary>
-        /// Queues a command to be sent to the IRC server. All prioritzed commands will be sent before non-prioritized commands.
+        /// Handles a connection alert and propogates it to any listeners.
         /// </summary>
-        public void SendCommand(string command, bool prioritized = false)
+        private void HandleConnectionAlert(ConnectionAlert alert)
         {
-            // Place command in respective queue
-            if (prioritized)
-                priorityOutputQueue.Enqueue(command);
-            else
-                outputQueue.Enqueue(command);
-        }
-
-        /// <summary>
-        /// Sends a chat message.
-        /// </summary>
-        public void SendChatMessage(string message)
-        {
-            if (message.Length <= 0) // Message can't be empty
-                return;
-
-            outputQueue.Enqueue("PRIVMSG #" + twitchDetails.channel + " :" + message); // Place message in queue
-        }
-
-        /// <summary>
-        /// Status types for the Twitch IRC client.
-        /// </summary>
-        public enum StatusType { Normal, Success, Error };
-
-        /// <summary>
-        /// Sends a connection state alert to the console and any listeners.
-        /// </summary>
-        private void ConnectionStateAlert(StatusType state, string message, int percentage = 0)
-        {
-            switch (state)
+            ConnectionAlertEvent?.Invoke(alert);
+            if (alert.isError)
             {
-                case StatusType.Error:
-                    Debug.LogError("<color=red><b>[ERROR]</b></color>: " + message);
-                    break;
-
-                case StatusType.Normal:
-                    Debug.Log("<color=#0018a1><b>[STATUS]</b></color>: <b>" + percentage + "%</b> " + message);
-                    break;
-
-                case StatusType.Success:
-                    Debug.Log("<color=#0ea300><b>[SUCCESS]</b></color>: <b>" + percentage + "%</b> " + message);
-                    break;
-            }
-
-            // Send status event to other listeners
-            StatusUpdateEvent?.Invoke(state, message, percentage);
-        }
-
-        #region Subclasses
-
-        [System.Serializable]
-        public class TwitchDetails
-        {
-            public string oauth = string.Empty;
-            public string nick = string.Empty;
-            public string channel = string.Empty;
-        }
-
-        [System.Serializable]
-        public class Settings
-        {
-            public bool connectOnStart = true;
-            public bool parseBadges = true;
-            public bool parseTwitchEmotes = true;
-            public bool debugIRC = true;
-        }
-
-        #endregion
-
-        private const int connectionCheckInterval = 500;
-        private Thread connectionThread = null;
-
-        /// <summary>
-        /// Thread process which checks whether the socket is still connected to the network.
-        /// </summary>
-        private void CheckSocketProc()
-        {
-            /* TODO
-             * Unclear why this is necessary. Sometimes, right after a new TcpClient is created, the socket says
-             * it has been shutdown. This catches that case and reconnects.
-            */
-
-            while (connected)
-            {
-                // Reconnect if the socket is disconnected
-                if (!CheckSocketConnection(client.Client))
-                {
-                    Debug.LogWarning("Socket is unexpectedly disconnected. Reconnecting...");
-                    taskQueue.Enqueue(() => Connect());
-                    break;
-                }
-                else
-                {
-                    Thread.Sleep(connectionCheckInterval);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Checks whether the socket is still connected to the network.
-        /// </summary>
-        private bool CheckSocketConnection(Socket socket)
-        {
-            var poll = socket.Poll(1000, SelectMode.SelectRead);
-            var avail = (socket.Available == 0);
-            if ((poll && avail) || !socket.Connected)
-            {
-                return false;
+                Debug.LogError(alert.message);
             }
             else
             {
-                return true;
+                Debug.Log(alert.message);
             }
+
+            // Reconnect if connection is interrupted
+            if (alert.status == ConnectionAlert.CONNECTION_INTERRUPTED)
+                Connect();
         }
     }